+++ outils/sync-démarches-simplifiées-88444.js
@@ -22,6 +22,7 @@ import _schema88444 from '../data/démarches-simplifiées/schema-DS-88444.json' 
 import {téléchargerNouveauxFichiersEspècesImpactées, téléchargerNouveauxFichiersAP_AM, téléchargerNouveauxFichiersMotivation} from './synchronisation-ds-88444/téléchargerNouveauxFichiersParType.js'
 import { ajouterDécisionsAdministratives, miseÀJourDécisionsAdministrativesDepuisDS88444 } from '../scripts/server/database/décision_administrative.js'
 import { déchiffrerDonnéesSupplémentairesDossiers } from '../scripts/server/démarches-simplifiées/chiffrerDéchiffrerDonnéesSupplémentaires.js'
+import { remplirChampsCommunsPourSynchro } from './synchronisation-ds-88444/remplirChampsCommunsPourSynchro.js'
 
 /** @import {default as DatabaseDossier} from '../scripts/types/database/public/Dossier.ts' */
 /** @import {default as Personne, PersonneInitializer} from '../scripts/types/database/public/Personne.ts' */
@@ -32,10 +33,10 @@ import { déchiffrerDonnéesSupplémentairesDossiers } from '../scripts/server/d
 
 /** @import {TypeDécisionAdministrative} from '../scripts/types/API_Pitchou.ts' */
 
-/** @import {DémarchesSimpliféesCommune, ChampDSCommunes, ChampDSDépartements, ChampDSRégions, Message, ChampDSDépartement, DémarchesSimpliféesDépartement, ChampScientifiqueIntervenants, BaseChampDS} from '../scripts/types/démarches-simplifiées/apiSchema.ts' */
-/** @import {DossierDS88444, Annotations88444, Champs88444} from '../scripts/types/démarches-simplifiées/apiSchema.ts' */
+/** @import {Message} from '../scripts/types/démarches-simplifiées/apiSchema.ts' */
+/** @import {DossierDS88444} from '../scripts/types/démarches-simplifiées/apiSchema.ts' */
 /** @import {SchemaDémarcheSimplifiée, ChampDescriptor} from '../scripts/types/démarches-simplifiées/schema.ts' */
-/** @import {DossierPourSynchronisation, DécisionAdministrativeAnnotation88444} from '../scripts/types/démarches-simplifiées/DossierPourSynchronisation.ts' */
+/** @import {DécisionAdministrativeAnnotation88444} from '../scripts/types/démarches-simplifiées/DossierPourSynchronisation.ts' */
 /** @import {DossierDemarcheSimplifiee88444, AnnotationsPriveesDemarcheSimplifiee88444} from '../scripts/types/démarches-simplifiées/DémarcheSimplifiée88444.ts' */
 /** @import {DonnéesSupplémentaires} from '../scripts/front-end/actions/importDossierUtils.js' */
 // récups les données de DS
@@ -128,390 +129,8 @@ const donnéesDécisionAdministrativeParNuméroDossier = new Map();
 
 
 
-/** @type {Omit<DossierPourSynchronisation, "demandeur_personne_physique" | "espèces_protégées_concernées" | "espèces_impactées" >[]} */
-const dossiersPourSynchronisation = dossiersDS.map((
-{
-    id: id_demarches_simplifiées,
-    number,
-    dateDepot: date_dépôt, 
-    demandeur,
-    champs,
-    annotations
-}) => {
-    /**
-     * Meta données
-     */
-    const number_demarches_simplifiées = String(number)
 
-    /*
-        Déposant 
-
-        Le déposant est la personne qui dépose le dossier sur DS
-        Dans certaines situations, cette personne est différente du demandeur (personne morale ou physique 
-        qui demande la dérogation), par exemple, si un bureau d'étude mandaté par une personne morale dépose 
-        le dossier
-        Le déposant n'est pas forcément représentant interne (point de contact principale) du demandeur
-
-        Dans la nomenclature DS, ce que nous appelons "déposant" se trouve dans la propriété "demandeur" 
-        (qui est différent de notre "demandeur")
-
-    */
-    /** @type {PersonneInitializer} */
-    let déposant;
-    {
-        const {prenom: prénoms, nom, email} = demandeur
-        déposant = {
-            prénoms,
-            nom,
-            email: email === '' ? undefined : email
-        }
-    }
-
-    /** 
-     * Champs 
-     */
-    /** @type {Map<string | undefined, Champs88444>} */
-    /** @type {Map<string | undefined, any>} */
-    const champById = new Map()
-    for(const champ of champs){
-        champById.set(champ.id, champ)
-    }
-
-    /** @type {DossierDemarcheSimplifiee88444['Nom du projet']} */ 
-    const nom = champById.get(pitchouKeyToChampDS.get('Nom du projet'))?.stringValue
-    /** @type {DossierDemarcheSimplifiee88444['Description synthétique du projet']} */ 
-    const description = champById.get(pitchouKeyToChampDS.get('Description synthétique du projet'))?.stringValue
-    /** @type {DossierDemarcheSimplifiee88444['Activité principale']} */ 
-    const activité_principale = champById.get(pitchouKeyToChampDS.get('Activité principale'))?.stringValue
-
-    /** @type {DossierDemarcheSimplifiee88444['Date de début d’intervention']} */ 
-    const date_début_intervention = champById.get(pitchouKeyToChampDS.get('Date de début d’intervention'))?.date
-    /** @type {DossierDemarcheSimplifiee88444['Date de fin d’intervention']} */ 
-    const date_fin_intervention = champById.get(pitchouKeyToChampDS.get('Date de fin d’intervention'))?.date
-    /** @type {DossierDemarcheSimplifiee88444['Durée de la dérogation']} */ 
-    const durée_intervention = Number(champById.get(pitchouKeyToChampDS.get('Durée de la dérogation'))?.stringValue)
-
-    /** @type {DossierDemarcheSimplifiee88444[`Synthèse des éléments démontrant qu'il n'existe aucune alternative au projet`]} */
-    const justification_absence_autre_solution_satisfaisante = champById.get(pitchouKeyToChampDS.get(`Synthèse des éléments démontrant qu'il n'existe aucune alternative au projet`))?.stringValue.trim()
-    /** @type {DossierDemarcheSimplifiee88444[`Motif de la dérogation`]} */
-    const motif_dérogation = champById.get(pitchouKeyToChampDS.get(`Motif de la dérogation`))?.stringValue
-        /** @type {DossierDemarcheSimplifiee88444[`Synthèse des éléments justifiant le motif de la dérogation`]} */
-    const justification_motif_dérogation = champById.get(pitchouKeyToChampDS.get(`Synthèse des éléments justifiant le motif de la dérogation`))?.stringValue.trim()
-
-
-
-
-    /* localisation */
-    /** @type {DossierDemarcheSimplifiee88444['Le projet se situe au niveau…'] | ''} */    
-    const projetSitué = champById.get(pitchouKeyToChampDS.get('Le projet se situe au niveau…'))?.stringValue
-    /** @type {ChampDSCommunes} */
-    const champCommunes = champById.get(pitchouKeyToChampDS.get('Commune(s) où se situe le projet'))
-    /** @type {ChampDSDépartements} */
-    const champDépartements = champById.get(pitchouKeyToChampDS.get('Département(s) où se situe le projet'))
-    /** @type {ChampDSDépartement} */
-    const champDépartementPrincipal = champById.get(pitchouKeyToChampDS.get('Dans quel département se localise majoritairement votre projet ?'))
-    /** @type {ChampDSRégions} */
-    const champRégions = champById.get(pitchouKeyToChampDS.get('Région(s) où se situe le projet'))
-
-
-    /** @type {DémarchesSimpliféesCommune[] | undefined} */
-    let communes;
-
-    /** @type {DémarchesSimpliféesDépartement['code'][] | undefined} */
-    let départements;
-    let régions;
-
-    if(projetSitué === `d'une ou plusieurs communes` && champCommunes){
-        communes = champCommunes.rows.map(c => c.champs[0].commune).filter(x => !!x)
-        
-        if(Array.isArray(communes) && communes.length >= 1){
-            départements = [...new Set(champCommunes.rows.map(c => c.champs[0].departement?.code).filter(x => !!x))]
-        }
-    }
-    else{
-        if(projetSitué === `d'un ou plusieurs départements` && champDépartements){
-            départements = [... new Set(champDépartements.rows.map(c => c.champs[0].departement?.code))]
-        }
-        else{
-            if(projetSitué === `d'une ou plusieurs régions` && champRégions){
-                régions = [... new Set(champRégions.rows.map(c => c.champs[0].stringValue))]
-            }
-            else{
-                if(projetSitué === 'de toute la France'){
-                    // ignorer
-                }
-                else{
-                    if(!projetSitué){
-                        // ignorer
-                    }
-                    else{
-                        console.log('localisation manquante', projetSitué, champs)
-                        process.exit(1)
-                    }
-                }
-            }
-        }
-    }
-
-    // Si la localisation avec les champs dédiés (surtout communes et départements) a échoué,
-    // se rabattre sur le champ du département principal s'il est présent
-    if(champDépartementPrincipal && champDépartementPrincipal.departement && (!départements || départements.length === 0)){
-        départements = [champDépartementPrincipal.departement.code]
-    }
-
-
-    /*
-        Demandeur
-     
-        Personne physique ou morale qui formule la demande de dérogation espèces protégées
-    */
-
-    /** @type {PersonneInitializer | undefined} */
-    /** let demandeur_personne_physique = undefined; */
-    /** @type {Entreprise | undefined} */
-    let demandeur_personne_morale = undefined
- 
-    const SIRETChamp = champById.get(pitchouKeyToChampDS.get('Numéro de SIRET'))
-    if(SIRETChamp){
-        const etablissement = SIRETChamp.etablissement
-        if(etablissement){
-            const { siret, address = {}, entreprise = {}} = etablissement
-            const {streetAddress, postalCode, cityName} = address
-            const {raisonSociale} = entreprise
-
-
-            demandeur_personne_morale = {
-                siret,
-                raison_sociale: raisonSociale,
-                adresse: `${streetAddress}\n${postalCode} ${cityName}`
-            }
-        }
-    }
-
-    /** Régime AE */
-    // le champ AE a changé d'une checkbox à un Oui/Non à un Oui/Non/Ne sait pas encore
-    // et donc, on gère les différentes valeurs pour les différentes version du formulaire
-
-    const rattaché_régime_ae_champ = champById.get(
-        pitchouKeyToChampDS.get("Le projet est-il soumis au régime de l'Autorisation Environnementale (article L. 181-1 du Code de l'environnement) ?")
-    )
-
-    const rattaché_au_régime_ae_stringValue = rattaché_régime_ae_champ?.stringValue
-
-    // null signifie "ne sait pas encore" et c'est la valeur par défaut
-    let rattaché_au_régime_ae = null;
-
-    if(rattaché_au_régime_ae_stringValue === 'Oui' || rattaché_au_régime_ae_stringValue === 'true'){
-        rattaché_au_régime_ae = true;
-    }
-    if(rattaché_au_régime_ae_stringValue === 'Non' || rattaché_au_régime_ae_stringValue === 'false'){
-        rattaché_au_régime_ae = false
-    }
-
-    /** Mesures ERC prévues */
-    const mesures_erc_prévues_champ = champById.get(
-        pitchouKeyToChampDS.get("Des mesures ERC sont-elles prévues ?")
-    )
-    const mesures_erc_prévues = mesures_erc_prévues_champ?.checked
-    
-    /** Données dossier scientifique */
-    /** @type {DossierDemarcheSimplifiee88444['Recherche scientifique - Votre demande concerne :']} */
-    const scientifique_type_demande_values = champById.get(pitchouKeyToChampDS.get('Recherche scientifique - Votre demande concerne :'))?.values
-    
-    /** @type {DossierDemarcheSimplifiee88444['Description du protocole de suivi']} */
-    const scientifique_description_protocole_suivi = champById.get(pitchouKeyToChampDS.get('Description du protocole de suivi'))?.stringValue
-
-    
-    /** @type {DossierDemarcheSimplifiee88444[`En cas de nécessité de capture d'individus, précisez le mode de capture`][]} */
-    const scientifique_précisez_mode_capture_values = champById.get(pitchouKeyToChampDS.get(`En cas de nécessité de capture d'individus, précisez le mode de capture`))?.values
-
-    /** @type {DossierDemarcheSimplifiee88444[`Préciser le(s) autre(s) moyen(s) de capture`]} */
-    const scientifique_precisez_autre_capture = champById.get(pitchouKeyToChampDS.get(`Préciser le(s) autre(s) moyen(s) de capture`))?.stringValue
-
-    /** @type {Set<DossierDemarcheSimplifiee88444[`En cas de nécessité de capture d'individus, précisez le mode de capture`] | DossierDemarcheSimplifiee88444[`Préciser le(s) autre(s) moyen(s) de capture`]>} */
-    const scientifique_mode_capture_set = scientifique_précisez_mode_capture_values ? new Set(scientifique_précisez_mode_capture_values) : new Set()
-    
-    if(scientifique_precisez_autre_capture){
-        scientifique_mode_capture_set.delete('Autre moyen de capture (préciser)')
-        scientifique_mode_capture_set.add(scientifique_precisez_autre_capture)
-    }
-
-    const scientifique_mode_capture = JSON.stringify([...scientifique_mode_capture_set])
-
-
-    /** @type {DossierDemarcheSimplifiee88444[`Utilisez-vous des sources lumineuses ?`]} */
-    const scientifique_modalités_source_lumineuses_boolean = champById.get(pitchouKeyToChampDS.get(`Utilisez-vous des sources lumineuses ?`))?.checked
-
-    /** @type {DossierDemarcheSimplifiee88444[`Précisez les modalités de l'utilisation des sources lumineuses`]} */
-    const scientifique_modalités_source_lumineuses_précisez = champById.get(pitchouKeyToChampDS.get(`Précisez les modalités de l'utilisation des sources lumineuses`))?.stringValue
-
-
-    const scientifique_modalités_source_lumineuses = scientifique_modalités_source_lumineuses_boolean && scientifique_modalités_source_lumineuses_précisez ?
-        scientifique_modalités_source_lumineuses_précisez : 
-        undefined
-
-    /** @type {DossierDemarcheSimplifiee88444[`Précisez les modalités de marquage pour chaque taxon`]} */
-    const scientifique_modalités_marquage = champById.get(pitchouKeyToChampDS.get(`Précisez les modalités de marquage pour chaque taxon`))?.stringValue || undefined
-
-
-    /** @type {DossierDemarcheSimplifiee88444[`Précisez les modalités de transport et la destination concernant la collecte de matériel biologique`]} */
-    const scientifique_modalités_transport = champById.get(pitchouKeyToChampDS.get(`Précisez les modalités de transport et la destination concernant la collecte de matériel biologique`))?.stringValue || undefined
-
-    /** @type {DossierDemarcheSimplifiee88444[`Précisez le périmètre d'intervention`]} */
-    const scientifique_périmètre_intervention = champById.get(pitchouKeyToChampDS.get(`Précisez le périmètre d'intervention`))?.stringValue || undefined
-
-    /** @type {ChampScientifiqueIntervenants | undefined} */
-    let scientifique_qualifications_intervenants = champById.get(pitchouKeyToChampDS.get(`Qualification des intervenants`)) || undefined
-
-    /** @type {BaseChampDS[][] | undefined} */
-    let rowsChamp = scientifique_qualifications_intervenants && 
-        scientifique_qualifications_intervenants.rows.map(r => r.champs)
-
-    /** @type { {nom_complet?: string, qualification?: string}[] | undefined} */
-    let scientifique_intervenants = undefined;
-
-    if(Array.isArray(rowsChamp)){
-        scientifique_intervenants = rowsChamp.map(champs => {
-            const champNomComplet = champs.find(c => c.label === 'Nom Prénom')
-            const champQualification = champs.find(c => c.label === 'Qualification')
-
-            return {
-                nom_complet: champNomComplet && champNomComplet.stringValue,
-                qualification: champQualification && champQualification.stringValue,
-            }
-        })
-    }
-
-    /** @type {DossierDemarcheSimplifiee88444[`Apporter des précisions complémentaires sur la possible intervention de stagiaire(s)/vacataire(s)/bénévole(s)`]} */
-    const scientifique_précisions_autres_intervenants = champById.get(pitchouKeyToChampDS.get(`Apporter des précisions complémentaires sur la possible intervention de stagiaire(s)/vacataire(s)/bénévole(s)`))?.stringValue || undefined;
-
-    /**
-     * Annotations privées
-     */
-    /** @type {Map<string | undefined, Annotations88444>} */
-    /** @type {Map<string | undefined, any>} */
-    const annotationById = new Map()
-    for(const annotation of annotations){
-        annotationById.set(annotation.id, annotation)
-    }
-
-    const historique_nom_porteur = annotationById.get(pitchouKeyToAnnotationDS.get("Nom du porteur de projet"))?.stringValue
-    const historique_localisation = annotationById.get(pitchouKeyToAnnotationDS.get("Localisation du projet"))?.stringValue
-    const ddep_nécessaire = annotationById.get(pitchouKeyToAnnotationDS.get("DDEP nécessaire ?"))?.stringValue
-
-    const enjeu_écologique = annotationById.get(pitchouKeyToAnnotationDS.get("Enjeu écologique")).checked
-    const enjeu_politique = annotationById.get(pitchouKeyToAnnotationDS.get("Enjeu politique")).checked
-    const commentaire_libre = annotationById.get(pitchouKeyToAnnotationDS.get("Commentaires sur les enjeux et la procédure")).stringValue
-
-    const historique_date_réception_ddep = annotationById.get(pitchouKeyToAnnotationDS.get("Date de réception DDEP")).date
-        
-    const historique_date_envoi_dernière_contribution = annotationById.get(pitchouKeyToAnnotationDS.get("Date d'envoi de la dernière contribution en lien avec l'instruction DDEP")).date
-    const historique_identifiant_demande_onagre = annotationById.get(pitchouKeyToAnnotationDS.get("N° Demande ONAGRE")).stringValue
-
-    const historique_date_saisine_csrpn = annotationById.get(pitchouKeyToAnnotationDS.get("Date saisine CSRPN")).date
-
-    const historique_date_saisine_cnpn = annotationById.get(pitchouKeyToAnnotationDS.get("Date saisine CNPN")) ?
-        annotationById.get(pitchouKeyToAnnotationDS.get("Date saisine CNPN")).date : 
-        undefined
-    
-    
-    const date_avis_csrpn = annotationById.get(pitchouKeyToAnnotationDS.get("Date avis CSRPN")).date
-    
-    const date_avis_cnpn = annotationById.get(pitchouKeyToAnnotationDS.get("Date avis CNPN")) ?
-        annotationById.get(pitchouKeyToAnnotationDS.get("Date avis CNPN")).date : 
-        undefined
-
-
-
-    const avis_csrpn_cnpn = annotationById.get(pitchouKeyToAnnotationDS.get("Avis CSRPN/CNPN")).stringValue
-
-    const date_consultation_public = annotationById.get(pitchouKeyToAnnotationDS.get("Date de début de la consultation du public ou enquête publique")).date
-
-    
-    const décision = annotationById.get(pitchouKeyToAnnotationDS.get("Décision")).stringValue
-    const date_signature_arrêté_préfectoral = annotationById.get(pitchouKeyToAnnotationDS.get("Date de signature de l'AP")).date
-    const référence_arrêté_préfectoral = annotationById.get(pitchouKeyToAnnotationDS.get("Référence de l'AP")).stringValue
-    const date_signature_arrêté_ministériel = annotationById.get(pitchouKeyToAnnotationDS.get("Date de l'AM")).date
-    const référence_arrêté_ministériel = annotationById.get(pitchouKeyToAnnotationDS.get("Référence de l'AM")).stringValue
-
-    donnéesDécisionAdministrativeParNuméroDossier.set(number_demarches_simplifiées, {
-        décision: décision || undefined,
-        date_signature_arrêté_préfectoral: date_signature_arrêté_préfectoral ? new Date(date_signature_arrêté_préfectoral) : undefined,
-        référence_arrêté_préfectoral: référence_arrêté_préfectoral || undefined,
-        date_signature_arrêté_ministériel: date_signature_arrêté_ministériel ? new Date(date_signature_arrêté_ministériel) : undefined,
-        référence_arrêté_ministériel: référence_arrêté_ministériel || undefined,
-    })
-
-    return {
-        // méta-données
-        id_demarches_simplifiées,
-        number_demarches_simplifiées,
-        date_dépôt,
-
-        // demandeur/déposant
-        // demandeur_personne_physique,
-        demandeur_personne_morale,
-        déposant,
-        //représentant,
-
-        // champs
-        nom,
-        description,
-        activité_principale,
-        date_début_intervention,
-        date_fin_intervention,
-        durée_intervention,
-
-        justification_absence_autre_solution_satisfaisante,
-        motif_dérogation,
-        justification_motif_dérogation,
-
-        // localisation
-        // https://knexjs.org/guide/schema-builder.html#json
-        communes: JSON.stringify(communes),
-        départements: JSON.stringify(départements),
-        régions: JSON.stringify(régions),
-
-        // régime AE
-        rattaché_au_régime_ae,
-
-        // mesurse ERC prévues
-        mesures_erc_prévues,
-
-        // données dossier scientifique
-        scientifique_type_demande: scientifique_type_demande_values ? JSON.stringify(scientifique_type_demande_values) : undefined,
-        scientifique_description_protocole_suivi,
-        scientifique_mode_capture,
-        scientifique_modalités_source_lumineuses,
-        scientifique_modalités_marquage,
-        scientifique_modalités_transport,
-        scientifique_périmètre_intervention,
-        scientifique_intervenants: JSON.stringify(scientifique_intervenants),
-        scientifique_précisions_autres_intervenants,
-
-        // annotations privées
-        historique_nom_porteur,
-        historique_localisation,
-        ddep_nécessaire,
-
-        enjeu_écologique,
-        enjeu_politique,
-        commentaire_libre,
-        
-        historique_date_réception_ddep,
-        historique_date_envoi_dernière_contribution,
-        historique_identifiant_demande_onagre,
-        historique_date_saisine_csrpn,
-        historique_date_saisine_cnpn,
-        date_avis_csrpn,
-        date_avis_cnpn,
-        avis_csrpn_cnpn,
-        
-        date_consultation_public,
-
-    }
-    
-})
+const dossiersPourSynchronisation = remplirChampsCommunsPourSynchro(dossiersDS, pitchouKeyToChampDS, pitchouKeyToAnnotationDS, donnéesDécisionAdministrativeParNuméroDossier)
